It builds now

+++ kernel/proc.c
@@ -798,6 +798,6 @@ int send_signal(int type, int sender_pid, int receiver_pid) {
   return 0;
 }
 
-int alarm(uint64 seconds) {
+int alarm(unsigned int seconds) {
   return 0;
 }
